Update journey_ecs.odin
Finished Really rough query and iteration. It works, I will heavily refactor and optimize it now. I will eventually implement sub-grouping optimization on the query in the future for the query functions. Created and commented With and Without function which can be used on the query to transform it (Not yet implemented).  Going to work the the TODO comment in the query section then refactor and optimize (excluding sub-group optimizing, since this will take a while)