Update journey_ecs.odin
added a better solution to creating entity, but this solution doesn't create entity with ID and Version(or Generation), but rather it only create entity with the ID. It save memory usage though by
Amount of Entity / 64. Going to think about if version and general meta data is important and be apart of the entity id. This solution will most likely change for better.