Update journey_ecs.odin
added removing all components from entity once it is deleted this implementation doesn't use the type remove, since we don't know all the component on entity, so created custom function just for it. Also added function to check memory usage a specific component sparse or all the component sparse. Removed len on Component sparse, and added meta_data struct on ComponentInfo struct which is needed.  Going to refactor and optimize the solution