Rough implementation on get_alive_entites
Added implementation to fetch all alive entities. Implementation is still rough, but it works. Going to flesh this out more. Then work on implementing  component iteration for per entity. Also improved the branching on internal_create_entity made it so branches are less likely taken and fall through which will result in a correct branch prediction till the branch is taken which will result on a miss predict refer to amd and intel documentation. This is on most common processor. To optimize for this in the developer side if you are removing entities do it in large bulks if possible.